requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.5.1
gspread==5.2.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
line-bot-sdk==2.1.0
MarkupSafe==2.1.1
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
